+++ rich_tools/__init__.py
@@ -1,3 +1,4 @@
+from .text import *
 from .table import *
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"